+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - backup-manager now uses a dedicated ServiceAccount, instead of the default one
 - service-binding controller now uses a dedicated ServiceAccount, instead of the default one
 - postgresql-operator now used a dedicated ServiceAccount, instead of the default one
+- backup custom resources now use a list of Conditions instead of a single enum as the status
 
 ## [0.1.0] - 2022-06-27